requirements.txt
python-telegram-bot==13.15
grpcio==1.50.0
protobuf==3.20.3
python-dotenv==0.19.2
PyYAML==6.0